Raster/requirements.txt
pillow
urllib3